requirements.txt
factual-api
flask
foursquare
fuzzywuzzy
gunicorn
pyrebase
pytest
pytz
pygeohash
python-dateutil
python-levenshtein
redis
rq
sanction
wikipedia
yelp
